pyproject.toml
[tool.poetry]
name = "nanoGPT-wrm"
version = "0.1.0"
description = "nanoGPT (forked)"
readme = "README.md"
packages = [{ include = "nanoGPT" }]
authors = ["W. Ross Morrow <morrowwr@gmail.com>"]

[[tool.poetry.source]]
name = 'default'
url = 'https://pypi.python.org/simple'
default = true

[tool.poetry.dependencies]
python  = "^3.9"
torch = "^2.0.0"
numpy = "^1.24.2"
transformers = "^4.27.1"
datasets = "^2.10.1"
tiktoken = "^0.3.2"
wandb = "^0.14.0"
tqdm = "^4.65.0"
pyyaml = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
black = "^22.6.0"
flake8 = "^4.0.1"
mypy = "^0.971"
types-pyyaml = "^6.0.12"

[tool.black]
line-length = 120
target-version = ['py39']

# TODO: this is a very loose setting, would be better to target
[tool.mypy]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
